Fix for proper handling mysql connections leftovers.

+++ aegir/tools/system/minute.sh
@@ -11,7 +11,7 @@ fi
 
 killit()
 {
-  if [ "$xtime" != "Time" ] && [ "$xuser" != "root" ] && [[ "$xtime" -gt "$limit" ]] ; then
+  if [ "$xtime" != "Time" ] && [ "$xuser" != "root" ] && [ "$xtime" != "|" ] && [[ "$xtime" -gt "$limit" ]] ; then
     xkill=`mysqladmin kill $each`
     times=`date`
     echo $times $each $xuser $xtime $xkill
@@ -26,6 +26,9 @@ xkill=null
 for each in `mysqladmin proc | awk '{print $2, $4, $8, $12}' | awk '{print $1}'`;
 do
   xtime=`mysqladmin proc | awk '{print $2, $4, $8, $12}' | grep $each | awk '{print $4}'`
+  if [ "$xtime" = "|" ] ; then
+    xtime=`mysqladmin proc | awk '{print $2, $4, $8, $11}' | grep $each | awk '{print $4}'`
+  fi
   xuser=`mysqladmin proc | awk '{print $2, $4, $8, $12}' | grep $each | awk '{print $2}'`
   if [ "$xtime" != "Time" ] ; then
     if [ "$xuser" = "xabuse" ] ; then
